[439871] Move the diagram refresh logic down to the non.ui plugin
Move the DiagramDialect and all the code it depends on in the non.ui
plugin.

Extract the ViewpointSelectionCallback logic in a dedicated class lower
in the stack (not in the UI) and delegating from the old one.

It is not compiling, we still have a reference to a
canonicalsynchronizerImpl and some logic to capture SWT line height.

Bug: 439871
Change-Id: I06f70d0264231ef364163f42cfbb49d2e6055bdb
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>