Introduce a ResourceMigrationMarker
This marker might be attached to a Resource to indicate that
some automatic migration *might* have change the structure
of the model during the loading (with no isModified flag)

This forces the IsModifiedSavingPolicy to consider this resource as
"logically modified".

This marker is automatically removed when the resource is processed
by the saving policy which checked that the serialization has no
difference with the current file content.

Bug: 430724
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr